Update sql.xml for Packages
Added Package and Package Body recognizing as nodes, very useful if you have both in the same file, or to go to package objects section (variables, constants and other declarations out of functions and procedures). I'm not yet able to put function and procedures under that node, but it's already better than nothing.